+++ src/mscorlib/shared/System.Private.CoreLib.Shared.projitems
@@ -340,6 +340,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)System\Security\UnverifiableCodeAttribute.cs"/>
     <Compile Include="$(MSBuildThisFileDirectory)System\Security\VerificationException.cs"/>
     <Compile Include="$(MSBuildThisFileDirectory)System\Span.cs"/>
+    <Compile Include="$(MSBuildThisFileDirectory)System\Span.NonGeneric.cs"/>
     <Compile Include="$(MSBuildThisFileDirectory)System\StackOverflowException.cs"/>
     <Compile Include="$(MSBuildThisFileDirectory)System\StringComparer.cs"/>
     <Compile Include="$(MSBuildThisFileDirectory)System\StringComparison.cs"/>
